# internal/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ScaledTest API",
    "description": "ScaledTest scales out end-to-end testing with unparalleled reporting and capabilities. This API provides CTRF report ingestion, distributed test execution, quality gate evaluation, webhook notifications, and analytics.",
    "version": "1.0.0",
    "license": { "name": "MIT" }
  },
  "servers": [{ "url": "http://localhost:8080", "description": "Local development" }],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "register",
        "summary": "Register a new user",
        "tags": ["Auth"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "Authenticate user",
        "tags": ["Auth"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "refreshToken",
        "summary": "Refresh access token",
        "description": "Uses the httpOnly refresh_token cookie to issue a new access token.",
        "tags": ["Auth"],
        "security": [],
        "responses": {
          "200": {
            "description": "Token refreshed",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/v1/reports": {
      "get": {
        "operationId": "listReports",
        "summary": "List test reports",
        "tags": ["Reports"],
        "parameters": [
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Offset" },
          {
            "name": "since",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "Filter reports created after this time"
          },
          {
            "name": "until",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "Filter reports created before this time"
          }
        ],
        "responses": {
          "200": {
            "description": "List of reports",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reports": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/TestReport" }
                    },
                    "total": { "type": "integer" }
                  },
                  "required": ["reports", "total"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "operationId": "createReport",
        "summary": "Submit a CTRF test report",
        "description": "Ingests a CTRF-formatted test report. Requires maintainer or owner role.",
        "tags": ["Reports"],
        "parameters": [
          {
            "name": "execution_id",
            "in": "query",
            "schema": { "type": "string", "format": "uuid" },
            "description": "Link report to an execution"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "CTRF report JSON (max 10MB)",
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CTRFReport" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "format": "uuid" },
                    "message": { "type": "string", "example": "report accepted" },
                    "tool": { "type": "string" },
                    "tests": { "type": "integer" },
                    "results": { "type": "integer" },
                    "execution_id": { "type": "string", "format": "uuid" },
                    "qualityGate": { "$ref": "#/components/schemas/QualityGateResult" }
                  },
                  "required": ["id", "message", "tool", "tests", "results"]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/v1/reports/compare": {
      "get": {
        "operationId": "compareReports",
        "summary": "Compare two reports",
        "tags": ["Reports"],
        "parameters": [
          {
            "name": "base",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "head",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Report comparison",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "base": { "$ref": "#/components/schemas/TestReport" },
                    "head": { "$ref": "#/components/schemas/TestReport" },
                    "diff": { "$ref": "#/components/schemas/ReportDiff" }
                  },
                  "required": ["base", "head", "diff"]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/reports/{reportID}": {
      "get": {
        "operationId": "getReport",
        "summary": "Get a single report",
        "tags": ["Reports"],
        "parameters": [
          {
            "name": "reportID",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Report details",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/TestReport" } }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "operationId": "deleteReport",
        "summary": "Delete a report",
        "tags": ["Reports"],
        "parameters": [
          {
            "name": "reportID",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Report deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "id": { "type": "string" }, "deleted": { "type": "boolean" } },
                  "required": ["id", "deleted"]
                }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/executions": {
      "get": {
        "operationId": "listExecutions",
        "summary": "List test executions",
        "tags": ["Executions"],
        "parameters": [
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Offset" }
        ],
        "responses": {
          "200": {
            "description": "List of executions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "executions": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/TestExecution" }
                    },
                    "total": { "type": "integer" }
                  },
                  "required": ["executions", "total"]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createExecution",
        "summary": "Create a test execution",
        "description": "Launches a distributed test execution. Requires maintainer or owner role.",
        "tags": ["Executions"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateExecutionRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Execution created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "format": "uuid" },
                    "status": { "type": "string", "example": "pending" },
                    "command": { "type": "string" }
                  },
                  "required": ["id", "status", "command"]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/v1/executions/{executionID}": {
      "get": {
        "operationId": "getExecution",
        "summary": "Get execution details",
        "tags": ["Executions"],
        "parameters": [
          {
            "name": "executionID",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Execution details",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/TestExecution" } }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "operationId": "cancelExecution",
        "summary": "Cancel an execution",
        "tags": ["Executions"],
        "parameters": [
          {
            "name": "executionID",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Execution cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "id": { "type": "string" }, "status": { "type": "string" } },
                  "required": ["id", "status"]
                }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/executions/{executionID}/status": {
      "put": {
        "operationId": "updateExecutionStatus",
        "summary": "Update execution status",
        "tags": ["Executions"],
        "parameters": [
          {
            "name": "executionID",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["running", "completed", "failed", "cancelled"]
                  },
                  "error_msg": { "type": "string" }
                },
                "required": ["status"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "id": { "type": "string" }, "status": { "type": "string" } },
                  "required": ["id", "status"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/executions/{executionID}/progress": {
      "post": {
        "operationId": "reportProgress",
        "summary": "Report execution progress",
        "tags": ["Executions"],
        "parameters": [
          {
            "name": "executionID",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "passed": { "type": "integer" },
                  "failed": { "type": "integer" },
                  "skipped": { "type": "integer" },
                  "total": { "type": "integer", "minimum": 1 },
                  "duration_ms": { "type": "integer" },
                  "estimated_eta_seconds": { "type": "number" }
                },
                "required": ["passed", "failed", "skipped", "total"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Progress received",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "execution_id": { "type": "string" },
                    "received": { "type": "boolean" }
                  },
                  "required": ["execution_id", "received"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/executions/{executionID}/test-result": {
      "post": {
        "operationId": "reportTestResult",
        "summary": "Report individual test result",
        "tags": ["Executions"],
        "parameters": [
          {
            "name": "executionID",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "status": {
                    "type": "string",
                    "enum": ["passed", "failed", "skipped", "pending", "other"]
                  },
                  "duration_ms": { "type": "integer" },
                  "message": { "type": "string" },
                  "suite": { "type": "string" },
                  "worker_id": { "type": "string" }
                },
                "required": ["name", "status"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result received",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "execution_id": { "type": "string" },
                    "received": { "type": "boolean" }
                  },
                  "required": ["execution_id", "received"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/executions/{executionID}/worker-status": {
      "post": {
        "operationId": "reportWorkerStatus",
        "summary": "Report worker status",
        "tags": ["Executions"],
        "parameters": [
          {
            "name": "executionID",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "worker_id": { "type": "string" },
                  "status": {
                    "type": "string",
                    "enum": ["starting", "running", "idle", "completed", "failed"]
                  },
                  "message": { "type": "string" },
                  "tests_assigned": { "type": "integer" },
                  "tests_completed": { "type": "integer" }
                },
                "required": ["worker_id", "status"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status received",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "execution_id": { "type": "string" },
                    "received": { "type": "boolean" }
                  },
                  "required": ["execution_id", "received"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/trends": {
      "get": {
        "operationId": "getAnalyticsTrends",
        "summary": "Pass rate trends over time",
        "tags": ["Analytics"],
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "Start time (default: 30 days ago)"
          },
          {
            "name": "end",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "End time (default: now)"
          },
          {
            "name": "group_by",
            "in": "query",
            "schema": { "type": "string", "enum": ["day", "week", "month"], "default": "day" }
          }
        ],
        "responses": {
          "200": {
            "description": "Trend data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "trends": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "date": { "type": "string", "format": "date-time" },
                          "total": { "type": "integer" },
                          "passed": { "type": "integer" },
                          "failed": { "type": "integer" },
                          "skipped": { "type": "integer" },
                          "pass_rate": { "type": "number" }
                        }
                      }
                    }
                  },
                  "required": ["trends"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/flaky-tests": {
      "get": {
        "operationId": "getFlakyTests",
        "summary": "Flaky test detection",
        "tags": ["Analytics"],
        "parameters": [
          { "name": "window_days", "in": "query", "schema": { "type": "integer", "default": 7 } },
          { "name": "min_runs", "in": "query", "schema": { "type": "integer", "default": 5 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Flaky tests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "flaky_tests": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "suite": { "type": "string" },
                          "file_path": { "type": "string" },
                          "flip_count": { "type": "integer" },
                          "total_runs": { "type": "integer" },
                          "flip_rate": { "type": "number" },
                          "last_status": { "type": "string" }
                        }
                      }
                    }
                  },
                  "required": ["flaky_tests"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/error-analysis": {
      "get": {
        "operationId": "getErrorAnalysis",
        "summary": "Error clustering and frequency",
        "tags": ["Analytics"],
        "parameters": [
          { "name": "start", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "end", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Error clusters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "message": { "type": "string" },
                          "count": { "type": "integer" },
                          "test_names": { "type": "array", "items": { "type": "string" } },
                          "first_seen": { "type": "string", "format": "date-time" },
                          "last_seen": { "type": "string", "format": "date-time" }
                        }
                      }
                    }
                  },
                  "required": ["errors"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/duration-distribution": {
      "get": {
        "operationId": "getDurationDistribution",
        "summary": "Test duration statistics",
        "tags": ["Analytics"],
        "parameters": [
          { "name": "start", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "end", "in": "query", "schema": { "type": "string", "format": "date-time" } }
        ],
        "responses": {
          "200": {
            "description": "Duration distribution",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "distribution": {
                      "type": "object",
                      "properties": {
                        "mean_ms": { "type": "number" },
                        "median_ms": { "type": "number" },
                        "p95_ms": { "type": "number" },
                        "p99_ms": { "type": "number" },
                        "min_ms": { "type": "integer" },
                        "max_ms": { "type": "integer" },
                        "distribution": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "range_label": { "type": "string" },
                              "min_ms": { "type": "integer" },
                              "max_ms": { "type": "integer" },
                              "count": { "type": "integer" }
                            }
                          }
                        }
                      }
                    }
                  },
                  "required": ["distribution"]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or sct_ API token",
        "description": "JWT access token (from /auth/login) or API token (sct_ prefix from team token creation)"
      }
    },
    "parameters": {
...
      "Offset": { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 8 },
          "display_name": { "type": "string", "minLength": 1 }
        },
        "required": ["email", "password", "display_name"]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        },
        "required": ["email", "password"]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/User" },
          "access_token": { "type": "string" },
          "expires_at": { "type": "string", "format": "date-time" }
        },
        "required": ["user", "access_token", "expires_at"]
      },
...
    }
  },
...
}